+++ GimmeMillions.Domain/Features/CandlestickStockFeatureExtractor.cs
@@ -16,11 +16,11 @@ namespace GimmeMillions.Domain.Features
 
         }
 
-        public float[] Extract(IEnumerable<(StockData Data, float Weight)> stocks)
+        public double[] Extract(IEnumerable<(StockData Data, float Weight)> stocks)
         {
             if(!stocks.Any())
             {
-                return new float[0];
+                return new double[0];
             }
 
             var ordered = stocks.OrderBy(x => x.Data.Date).ToList();
@@ -28,15 +28,15 @@ namespace GimmeMillions.Domain.Features
 
             decimal average = (lastStock.Data.Close + lastStock.Data.Open + lastStock.Data.Low + lastStock.Data.High) / 4.0m;
             decimal averageVolume = ordered.Average(x => x.Data.Volume);
-            var feature = new float[stocks.Count() * 5];
+            var feature = new double[stocks.Count() * 5];
             int index = 0;
             foreach(var stock in ordered)
             {
-                feature[index * 5] = (float)(stock.Data.Open - average);
-                feature[index * 5 + 1] = (float)(stock.Data.Close - average);
-                feature[index * 5 + 2] = (float)(stock.Data.High - average);
-                feature[index * 5 + 3] = (float)(stock.Data.Low - average);
-                feature[index * 5 + 4] = (float)(stock.Data.Volume / averageVolume);
+                feature[index * 5] = (double)(stock.Data.Open - average);
+                feature[index * 5 + 1] = (double)(stock.Data.Close - average);
+                feature[index * 5 + 2] = (double)(stock.Data.High - average);
+                feature[index * 5 + 3] = (double)(stock.Data.Low - average);
+                feature[index * 5 + 4] = (double)(stock.Data.Volume / averageVolume);
 
                 index++;
             }
